Move the aws utils
